requirements.txt
Django==4.2.*
Pillow==10.4.*
bleach==6.1.*
pytest==8.*
pytest-django==4.*
django-widget-tweaks==1.5.*
django-allauth==0.63.*
